Prevent error in stat.php and list.php
The error was caused by invalid access to unset values of $_GET.
We have not cared about the error because the error log is the only
place it would be reported to, and the scripts themselves seem to work
properly.  But we'd like to fix it because the error log would be
cluttered with this error, which could hide really problematic errors.